Platform.App/python-template/requirements.txt
alembic==0.9.7
astroid==1.6.1
attrs==17.3.0
autopep8==1.3.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
coverage==4.4.2
decorator==4.2.1
docker==2.7.0
docker-pycreds==0.2.1
falcon==1.3.0
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
json-rpc==1.10.8
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
parso==0.1.1
pbr==3.1.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
py==1.5.2
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pytest==3.3.1
pytest-cov==2.5.1
pytest-flask==0.10.0
python-dateutil==2.6.1
python-editor==1.0.3
python-language-server==0.14.0
python-mimeparse==1.6.0
PyYAML==3.12
requests==2.18.4
rope==0.10.7
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
SQLAlchemy==1.2.2
SQLAlchemy-Utils==0.33.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.20.2
